Refactored the getSysUserId Method into a Single Class
