+++ src/main/java/org/openelisglobal/common/rest/DisplayListController.java
@@ -28,6 +28,7 @@ import org.openelisglobal.common.services.StatusService;
 import org.openelisglobal.common.services.StatusService.AnalysisStatus;
 import org.openelisglobal.common.util.ConfigurationProperties;
 import org.openelisglobal.common.util.ConfigurationProperties.Property;
+import org.openelisglobal.common.util.ControllerUtills;
 import org.openelisglobal.common.util.DateUtil;
 import org.openelisglobal.common.util.IdValuePair;
 import org.openelisglobal.common.util.LabelValuePair;
@@ -73,7 +74,7 @@ import org.springframework.web.bind.annotation.ResponseBody;
 
 @Controller
 @RequestMapping(value = "/rest/")
-public class DisplayListController extends BaseRestController {
+public class DisplayListController extends ControllerUtills {
     @Value("${org.itech.login.saml:false}")
     private Boolean useSAML;
