+++ src/main/java/org/openelisglobal/common/rest/provider/AllTestsForSampleTypeProviderRestController.java
@@ -20,7 +20,7 @@ import java.util.List;
 import org.json.simple.JSONArray;
 import org.json.simple.JSONObject;
 import org.openelisglobal.common.log.LogEvent;
-import org.openelisglobal.common.rest.BaseRestController;
+import org.openelisglobal.common.util.ControllerUtills;
 import org.openelisglobal.spring.util.SpringContext;
 import org.openelisglobal.test.valueholder.Test;
 import org.openelisglobal.typeofsample.service.TypeOfSampleService;
@@ -33,7 +33,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequestMapping("/rest")
-public class AllTestsForSampleTypeProviderRestController extends BaseRestController {
+public class AllTestsForSampleTypeProviderRestController extends ControllerUtills {
 
     private TypeOfSampleService typeOfSampleService = SpringContext.getBean(TypeOfSampleService.class);
