+++ src/main/java/org/openelisglobal/qaevent/controller/rest/NonConformingEventsCorrectionActionRestController.java
@@ -8,7 +8,7 @@ import java.util.Map;
 import javax.servlet.http.HttpServletRequest;
 import org.apache.commons.validator.GenericValidator;
 import org.openelisglobal.common.services.DisplayListService;
-import org.openelisglobal.login.valueholder.UserSessionData;
+import org.openelisglobal.common.util.ControllerUtills;
 import org.openelisglobal.patient.action.bean.PatientSearch;
 import org.openelisglobal.qaevent.form.NonConformingEventForm;
 import org.openelisglobal.qaevent.service.NCEventService;
@@ -26,11 +26,9 @@ import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequestMapping(value = "/rest")
-public class NonConformingEventsCorrectionActionRestController {
+public class NonConformingEventsCorrectionActionRestController extends ControllerUtills {
 
     private NCEventService ncEventService = SpringContext.getBean(NCEventService.class);
-
-    private static final String USER_SESSION_DATA = "userSessionData";
 
     @Autowired
     private NonConformingEventWorker nonConformingEventWorker;
@@ -89,16 +87,5 @@ public class NonConformingEventsCorrectionActionRestController {
         } else {
             return ResponseEntity.ok().body(Map.of("success", false));
         }
-    }
-
-    protected String getSysUserId(HttpServletRequest request) {
-        UserSessionData usd = (UserSessionData) request.getSession().getAttribute(USER_SESSION_DATA);
-        if (usd == null) {
-            usd = (UserSessionData) request.getAttribute(USER_SESSION_DATA);
-            if (usd == null) {
-                return null;
-            }
-        }
-        return String.valueOf(usd.getSystemUserId());
     }
 }
