+++ src/main/java/org/openelisglobal/qaevent/controller/rest/ReportNonConformEventsRestController.java
@@ -9,8 +9,8 @@ import org.openelisglobal.common.rest.bean.NceSampleInfo;
 import org.openelisglobal.common.rest.bean.NceSampleItemInfo;
 import org.openelisglobal.common.services.DisplayListService;
 import org.openelisglobal.common.services.RequesterService;
+import org.openelisglobal.common.util.ControllerUtills;
 import org.openelisglobal.common.util.DateUtil;
-import org.openelisglobal.login.valueholder.UserSessionData;
 import org.openelisglobal.qaevent.form.NonConformingEventForm;
 import org.openelisglobal.qaevent.service.NceCategoryService;
 import org.openelisglobal.qaevent.valueholder.NcEvent;
@@ -33,7 +33,7 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-public class ReportNonConformEventsRestController {
+public class ReportNonConformEventsRestController extends ControllerUtills {
 
     private final SampleService sampleService;
     private final SampleItemService sampleItemService;
@@ -41,8 +41,6 @@ public class ReportNonConformEventsRestController {
     private final NonConformingEventWorker nonConformingEventWorker;
     private final NceCategoryService nceCategoryService;
     private final RequesterService requesterService;
-
-    private static final String USER_SESSION_DATA = "userSessionData";
 
     @Autowired
     private SystemUserService systemUserService;
@@ -173,17 +171,6 @@ public class ReportNonConformEventsRestController {
 
         sampleInfo.setSampleItems(sampleItemsList);
         return sampleInfo;
-    }
-
-    protected String getSysUserId(HttpServletRequest request) {
-        UserSessionData usd = (UserSessionData) request.getSession().getAttribute(USER_SESSION_DATA);
-        if (usd == null) {
-            usd = (UserSessionData) request.getAttribute(USER_SESSION_DATA);
-            if (usd == null) {
-                return null;
-            }
-        }
-        return String.valueOf(usd.getSystemUserId());
     }
 
     private Sample getSampleForLabNumber(String labNumber) throws LIMSInvalidConfigurationException {
