+++ src/main/java/org/openelisglobal/reports/controller/rest/ReportRestController.java
@@ -16,7 +16,7 @@ import javax.servlet.http.HttpServletResponse;
 import net.sf.jasperreports.engine.JRException;
 import org.openelisglobal.common.exception.LIMSRuntimeException;
 import org.openelisglobal.common.log.LogEvent;
-import org.openelisglobal.common.rest.BaseRestController;
+import org.openelisglobal.common.util.ControllerUtills;
 import org.openelisglobal.common.util.validator.GenericValidator;
 import org.openelisglobal.reports.action.implementation.IReportCreator;
 import org.openelisglobal.reports.action.implementation.ReportImplementationFactory;
@@ -30,7 +30,7 @@ import org.springframework.web.bind.annotation.ResponseBody;
 
 @Controller
 @RequestMapping(value = "/rest/")
-public class ReportRestController extends BaseRestController {
+public class ReportRestController extends ControllerUtills {
 
     @Autowired
     private ServletContext context;
