+++ src/main/java/org/openelisglobal/sample/service/PatientManagementUpdate.java
@@ -16,6 +16,7 @@ import org.openelisglobal.address.valueholder.PersonAddress;
 import org.openelisglobal.common.action.IActionConstants;
 import org.openelisglobal.common.exception.LIMSRuntimeException;
 import org.openelisglobal.common.log.LogEvent;
+import org.openelisglobal.common.util.ControllerUtills;
 import org.openelisglobal.common.validator.BaseErrors;
 import org.openelisglobal.login.valueholder.UserSessionData;
 import org.openelisglobal.patient.action.IPatientUpdate;
@@ -39,7 +40,7 @@ import org.springframework.validation.Errors;
 
 @Service
 @Scope("prototype")
-public class PatientManagementUpdate implements IPatientUpdate {
+public class PatientManagementUpdate extends ControllerUtills implements IPatientUpdate {
 
     private String currentUserId;
     protected Patient patient;
@@ -76,11 +77,6 @@ public class PatientManagementUpdate implements IPatientUpdate {
                 ADDRESS_PART_VILLAGE_ID = addressPart.getId();
             }
         }
-    }
-
-    protected String getSysUserId(HttpServletRequest request) {
-        UserSessionData usd = (UserSessionData) request.getSession().getAttribute(IActionConstants.USER_SESSION_DATA);
-        return String.valueOf(usd.getSystemUserId());
     }
 
     public void setSysUserIdFromRequest(HttpServletRequest request) {
