+++ src/main/java/org/openelisglobal/workplan/controller/rest/PrintWorkplanReportRestController.java
@@ -15,7 +15,7 @@ import net.sf.jasperreports.engine.JasperRunManager;
 import net.sf.jasperreports.engine.data.JRBeanCollectionDataSource;
 import org.openelisglobal.common.exception.LIMSRuntimeException;
 import org.openelisglobal.common.log.LogEvent;
-import org.openelisglobal.common.rest.BaseRestController;
+import org.openelisglobal.common.util.ControllerUtills;
 import org.openelisglobal.test.service.TestServiceImpl;
 import org.openelisglobal.workplan.form.WorkplanForm;
 import org.openelisglobal.workplan.form.WorkplanForm.PrintWorkplan;
@@ -29,7 +29,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController("PrintWorkplanReportRestController")
-public class PrintWorkplanReportRestController extends BaseRestController {
+public class PrintWorkplanReportRestController extends ControllerUtills {
 
     private String reportPath = null;
